dashboard-service/requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
celery[redis]
fastapi[celery]
requests
